+++ bot_modules/__init__.py
@@ -1,0 +1,4 @@
+# version 1.3.0
+from .snapshots import create_snapshot, list_snapshots, delete_snapshot, restore_snapshot, download_snapshot
+
+__all__ = ['create_snapshot', 'list_snapshots', 'delete_snapshot', 'restore_snapshot', 'download_snapshot']
